pyproject.toml
[tool.poetry]
name = "lpms-self-localization"
version = "0.1.0"
description = ""
authors = ["dragoon8192 <dragoon8192@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
pandas = "^1.4.1"
numpy = ">=1.18, <1.22"
scipy = "^1.8.0"
numpy-quaternion = "^2022.2.10"
numba = "^0.55.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-language-server = "^0.36.2"

[tool.poetry.scripts]
app = "lpms_self_localization:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
